requirements.txt
streamlit
openai
langchain-openai
requests
python-dotenv